+++ services/IAuctionEntryService.cs
@@ -6,7 +6,7 @@ public interface IAuctionEntryService
 {
     public Task<bool> CreateAuctionEntry(string auctionId, AuctionEntry auctionEntryService);
 
-    public Task<int> UpdateSoldAuctionEntries(string previousAuctionId, List<long> soldAuctionEntries);
+    public Task<int> UpdateSoldAuctionEntries(string previousAuctionId, List<int> soldAuctionEntries);
 
     public Task<List<AuctionEntry>> GetAuctionEntriesFromAuctionId(string auctionId);
 }
